Time Series Analysis using R programming.
ARIMA model focuses on analyzing time series linearly and it does not reflect recent changes as new information is available. Therefore, in order to update the model, users need to incorporate new data and estimate parameters again. The variance in ARIMA model is unconditional variance and remains constant. ARIMA is applied for stationary series and therefore, non-stationary series should be transformed (such as log transformation).
Additionally, ARIMA is often used together with ARCH/GARCH model. ARCH/GARCH is a method to measure volatility of the series, or more specifically, to model the noise term of ARIMA model. ARCH/GARCH incorporates new information and analyzes the series based on conditional variances where users can forecast future values with up-to-date information. The forecast interval for the combined (mixed) model is closer than that of fitted ARIMA model.
It is to be remembered that due to the presences of high volatility in the stock market prices resulting in the Heteroscedastic (Varying Variance) behavior of the residuals, the ARIMA and GARCH models can only be used mainly for Short Term Forecasting of the stock prices. It might not given reliable results in case of long term forecasting of the stock prices due to the presence of high Volatility in the stock markets.
Thus, in short, we transferred the whole Nifty 50 data from Excel to R platform whose layout is given in the Table 3.1. We extracted the Adjusted Closing Prices from the original data, cleaned them and constructed the plot of the extracted Adjusted Closing values given in Diagram 3.1. We checked for the Stationarity of the Adjusted Closing Prices and found that it is non – stationary given in Table 3.2.We did suitable Transformations (For e.g. Log transformation, Differencing. etc.) to the data to make it Stationary and plotted them, given in Diagram 3.2 and interpreted them.
	

After this, we proceeded towards the ARIMA Model Fitting of the Prices. We identified the model using the Auto - Correlation F unction (ACF) and Partial Auto - Correlation Function (PACF) plots, given in Diagram 3.3 and Akaike Information Criteria (AIC) in Table 3.4.We found that the ARIMA (1, 1, 1) model with drift has the least AIC and thus we choose this model. Next, we estimated the parameters and also found there corresponding standard errors, 2 values and p -values to test their individual significance given in Table 3.5.Then, we tried to forecast the future prices for the upcoming 100 days and the log forecast plot of the log prices and we also provided the different forecasting errors provided in Table 3.6. We also compared the fitted prices with the original prices in Diagram 3.5. To check the validity of the model, we proceeded towards the diagnostic checking of the residuals of the fitted ARIMA model. We provided the Residuals plot in Diagram 3.6 to check the pattern of the residuals, ACF and PACF plots to study the Auto Correlation structure of the residuals in Diagram 3.7and also constructed the Q - Q plots of the residuals and the squared residuals to test their normality in Diagram 3.8.Finally, we performed the Lj ung - Box test of the squared residuals in Table 3.8 to conclude that there is a trace of conditional heteroscedasticity in the residuals of the fitted model.

Thus, the presence of volatility in the concerned time series became inevitable and to be sure, we further studied the Returns, the Squared Returns plots given in Diagram 3.10and also the ACF and the PACF plots of the Squared Returns in Diagram 3.11.We, then, fitted the GARCH model on the returns of the fitted ARIMA model using the values of the MOS for different types of model, given in Table 3.9. We saw that GARCH (1, 1) model had the least AIC. Thus, we estimated the parameters of the GARCH model and also provided their corresponding standard errors, t- values and p- values to check their significance Table 3.10.We plotted the Conditional Standard deviations and the Conditional Variances in Diagram 3.12.


We combined the two fitted models, that is the previously fitted ARIMA (1, 1, 1) model and the fitted GACRH (0, 2) model. The model is given below.
The combined ARIMA (1, 1, 1) + GARCH (0, 2) model is given as,
Yt-Yt-1= 4.3366 *10-4 – 0.57969632(Yt-1 –Yt-2 ) + 0.65876704ɛt-1 1.113e-08  + 2.506e-02* ɛ2t-1                  +2.955e+00*ht-1

After this, we proceeded to compare the two models. Using the previously fitted ARIMA model and the above combined model for predicting the stock prices, we calculated the 1 - step forecast values along with their 95% CIs given in Table 3.11 and Table 3.12. We clearly found that the 95% CI in case of the former model was much wider than that of the latter combined model. Thus, the combined model gives a good forecast model as here. The volatility is also taken care of apart from modelling the original prices only. We also plotted the fitted stock prices along with their 95% using the combined fitted model Diagram 3.14. At last, we did the diagnostic checking of the returns of the combined model. We compared the Q Q plot of the returns of the combined model with that of the fitted ARIMA model Diagram. 3.15.After performing the Ljung 'Box test on the residuals of the combined model, provided in Table 3.13, we claimed that as we have successfully modelled the heteroscedastic nature of the returns of the ARIMA model and thus, the combined model is has Performed better than the previously fitted ARIMA model. Thus, it can be used for short –term forecasting of the Adjusted Closing Stock Prices.